requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
requests==2.32.3
python-dateutil==2.9.0.post0
psycopg[binary]==3.2.1
transformers==4.44.2
torch==2.4.1
